Add initial version of a plugin-framework provider config test, move code between packages to allow tests (#8797)
* Add initial version of plugin framework provider config test affected by inaccessible functions

* Move `ReplaceVarsForFrameworkTest` to `fwresource` package

This mirrors how old, ReplaceVars* functions are in the `tpgresource` package

* Move functions for changing ENVs in tests to `acctest` package, update usage

* Move `GenerateFakeCredentialsJson` function to `acctest` package, update usage

* Refactor provider config tests to use plugin-framework types

* Add all test cases to match equivalent SDK provider config tests

Some test cases currently fail due to different handling of empty values by the SDK and plugin framework

* Add test case about handling of Unknown values for `project`

* Update tests to check values in BOTH the data model and provider config struct after `LoadAndValidateFramework` runs
