+++ google-beta/acctest/framework_test_utils.go
@@ -6,18 +6,12 @@ import (
 	"context"
 	"fmt"
 	"log"
-	"reflect"
-	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/fwtransport"
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
 )
 
 func GetFwTestProvider(t *testing.T) *frameworkTestProvider {
@@ -80,49 +74,4 @@ func testStringValue(sPtr *string) string {
 	}
 
 	return *sPtr
-}
-
-// This function isn't a test of transport.go; instead, it is used as an alternative
-// to ReplaceVars inside tests.
-func ReplaceVarsForFrameworkTest(prov *fwtransport.FrameworkProviderConfig, rs *terraform.ResourceState, linkTmpl string) (string, error) {
-	re := regexp.MustCompile("{{([[:word:]]+)}}")
-	var project, region, zone string
-
-	if strings.Contains(linkTmpl, "{{project}}") {
-		project = rs.Primary.Attributes["project"]
-	}
-
-	if strings.Contains(linkTmpl, "{{region}}") {
-		region = tpgresource.GetResourceNameFromSelfLink(rs.Primary.Attributes["region"])
-	}
-
-	if strings.Contains(linkTmpl, "{{zone}}") {
-		zone = tpgresource.GetResourceNameFromSelfLink(rs.Primary.Attributes["zone"])
-	}
-
-	replaceFunc := func(s string) string {
-		m := re.FindStringSubmatch(s)[1]
-		if m == "project" {
-			return project
-		}
-		if m == "region" {
-			return region
-		}
-		if m == "zone" {
-			return zone
-		}
-
-		if v, ok := rs.Primary.Attributes[m]; ok {
-			return v
-		}
-
-		// Attempt to draw values from the provider
-		if f := reflect.Indirect(reflect.ValueOf(prov)).FieldByName(m); f.IsValid() {
-			return f.String()
-		}
-
-		return ""
-	}
-
-	return re.ReplaceAllStringFunc(linkTmpl, replaceFunc), nil
 }
